refactor(router): move route_handlers into rules folder
