+++ modules/angular2/src/router/rules/route_handlers/async_route_handler.ts
