+++ modules/angular2/src/router/rules/route_handlers/route_handler.ts
