+++ modules/angular2/src/router/rules/route_handlers/sync_route_handler.ts
