+++ modules/angular2/src/router/rules/rules.ts
@@ -3,7 +3,7 @@ import {BaseException} from 'angular2/src/facade/exceptions';
 import {PromiseWrapper} from 'angular2/src/facade/promise';
 import {Map} from 'angular2/src/facade/collection';
 
-import {RouteHandler} from '../route_handlers/route_handler';
+import {RouteHandler} from './route_handlers/route_handler';
 import {Url, serializeParams} from './url_parser';
 import {ComponentInstruction} from '../instruction';
 import {ParamRoutePath} from './param_route_path';
